refactor: move provider_errorcases_test.go to internal/acctest/
TestUnit* tests in provider_errorcases_test.go use resource.UnitTest
(no TF_ACC required) but depend on the provider factory and the embedded
Terraform binary — they test schema-level validation visible to users,
not internal implementation details.

Changes:
- internal/provider/provider_errorcases_test.go deleted
- internal/acctest/errorcases_test.go created (package acctest,
  testAccProtoV6ProviderFactories -> ProtoV6ProviderFactories)
- internal/provider/provider_test.go deleted — the bridge variable
  testAccProtoV6ProviderFactories had no remaining consumers in
  internal/provider/ after this move

go test ./internal/provider/... -> ok (unit tests only, no credentials)
go test ./internal/acctest/...  -> ok (TestUnit* run; TestAcc* skipped)

Closes #298
